+++ ui/panels-plugin/src/plugins/time-series-chart/TimeSeriesChartPanel.tsx
@@ -19,7 +19,7 @@ import { Box, Skeleton } from '@mui/material';
 import { LineChart, EChartsDataFormat, ZoomEventData, Legend } from '@perses-dev/components';
 import { useSuggestedStepMs } from '../../model/time';
 import { StepOptions, ThresholdColors, ThresholdColorsPalette } from '../../model/thresholds';
-import { TimeSeriesChartOptions, DEFAULT_LEGEND, DEFAULT_UNIT } from './time-series-chart-model';
+import { TimeSeriesChartOptions, DEFAULT_LEGEND, DEFAULT_UNIT, DEFAULT_VISUAL } from './time-series-chart-model';
 import {
   getLineSeries,
   getThresholdSeries,
@@ -42,6 +42,10 @@ export function TimeSeriesChartPanel(props: TimeSeriesChartProps) {
   const legend = props.spec.legend ? merge({}, DEFAULT_LEGEND, props.spec.legend) : undefined;
 
   const unit = props.spec.unit ?? DEFAULT_UNIT;
+
+  // ensures there are fallbacks for unset properties since most
+  // users should not need to customize visual display
+  const visual = merge({}, DEFAULT_VISUAL, props.spec.visual);
 
   // TODO: change to array, support multi select on Shift-click
   const [selectedSeriesName, setSelectedSeriesName] = useState<string | null>(null);
@@ -90,7 +94,7 @@ export function TimeSeriesChartPanel(props: TimeSeriesChartProps) {
       for (const timeSeries of query.data.series) {
         const formattedSeriesName = timeSeries.formattedName ?? timeSeries.name;
         const yValues = getYValues(timeSeries, timeScale);
-        const lineSeries = getLineSeries(timeSeries.name, formattedSeriesName, yValues, selectedSeriesName);
+        const lineSeries = getLineSeries(timeSeries.name, formattedSeriesName, yValues, visual);
         if (selectedSeriesName === null || selectedSeriesName === timeSeries.name) {
           graphData.timeSeries.push(lineSeries);
         }
@@ -130,7 +134,7 @@ export function TimeSeriesChartPanel(props: TimeSeriesChartProps) {
       loading: queriesFinished === 0,
       allQueriesLoaded: queriesFinished === queryResults.length,
     };
-  }, [queryResults, thresholds, selectedSeriesName, legend]);
+  }, [queryResults, thresholds, selectedSeriesName, legend, visual]);
 
   if (contentDimensions === undefined) {
     return null;